Adding python dependencies installation and improving data base installation method

+++ src/build/build.py
@@ -1,18 +1,28 @@
 import argparse
 import os
 import getpass
-import logging
 import re
 import subprocess
 import signal
 import sys
 from time import gmtime, strftime
-import pymysql.cursors
+from src.tools import logger
 
 def interrupt_handler (exception_info):
-	logging.debug('Interrupt signal recieved: %s', str(exception_info))
-	logging.warning('Stopping installation.')
+	build_logger.debug('Interrupt signal recieved: ' + str(exception_info))
+	build_logger.warning('Stopping installation.')
 	exit(2)
+
+def create_logger ():
+	build_logger = logger.acsLogger()
+	if args.debug:
+		build_logger.set_debug()
+	elif args.verbose:
+		build_logger.set_info()
+	else:
+		build_logger.set_warning()
+
+	return build_logger
 
 def arg_parser():
 	parser = argparse.ArgumentParser(description='Installation script for ACS.')
@@ -32,7 +42,7 @@ def _get_system_architecture ():
 def _is_architecture_supported ():
 	SUPPORTED_ARCHITECTURES = ['x86_64', 'armvl7']
 	my_architecture = _get_system_architecture()
-	logging.info('System architecture is %s', my_architecture)
+	build_logger.info('System architecture is ' + my_architecture)
 	if my_architecture in SUPPORTED_ARCHITECTURES:
 		return True
 	else:
@@ -55,7 +65,7 @@ def _command_list_to_str(listed_commands):
 	result = separator.join(listed_commands)
 	return result
 
-def _install_packages (args, list_of_packages, stacked_call=False):
+def _install_packages (list_of_packages, stacked_call=False):
 	if list_of_packages is not None:
 		if args.debug:
 			installation_cmd = ['sudo', 'apt-get', 'install', '-y']
@@ -63,33 +73,33 @@ def _install_packages (args, list_of_packages, stacked_call=False):
 			installation_cmd = ['sudo', 'apt-get', '-qq', 'install', '-y']
 		installation_cmd.extend(list_of_packages)
 		if not stacked_call:
-			logging.info('Installing necessary packages: %s', _command_list_to_str(installation_cmd))
+			build_logger.info('Installing necessary packages: ' + _command_list_to_str(installation_cmd))
 		else:
-			logging.info('Retrying installation of necessary packages: %s', _command_list_to_str(installation_cmd))
+			build_logger.info('Retrying installation of necessary packages: ' + _command_list_to_str(installation_cmd))
 		try:
 			subprocess.run(installation_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True)
 		except KeyboardInterrupt:
 			interrupt_handler(sys.exc_info())
 		except subprocess.CalledProcessError as installation_error:
 			if not stacked_call:
-				logging.warning ('Could not install packages. Updating apt.')
-				_update_apt(args)
-				_install_packages(args, list_of_packages, True)
+				build_logger.warning ('Could not install packages. Updating apt.')
+				_update_apt()
+				_install_packages(list_of_packages, True)
 			else:
-				logging.error('Something went wrong when trying to install packages.')
-				logging.error('Dumping apt error message\n\t%s', installation_error.stderr.decode('utf-8'))
+				build_logger.error('Something went wrong when trying to install packages.')
+				build_logger.error('Dumping apt error message\n\t ' + installation_error.stderr.decode('utf-8'))
 				exit (1)
-		logging.info('Package installation was successfull!')
+		build_logger.info('Package installation was successfull!')
 	else:
-		logging.error('Could not get list of packages.')
+		build_logger.error('Could not get list of packages.')
 		exit (1)
 
-def _update_apt (args):
+def _update_apt ():
 	if args.debug:
 		apt_update_cmd = ['sudo', 'apt-get', 'update', '-y']
 	else:
 		apt_update_cmd = ['sudo', 'apt-get', '-qq', 'update', '-y']
-	logging.debug('Updating apt: %s', _command_list_to_str(apt_update_cmd))
+	build_logger.debug('Updating apt: ' + _command_list_to_str(apt_update_cmd))
 	try:
 		subprocess.run(apt_update_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True)
 	except KeyboardInterrupt:
@@ -104,35 +114,34 @@ def _call_std_subprocess (command):
 		print_process_error(err)
 
 def print_process_error (process):
-		logging.error(process.stderr.decode('utf-8'))
+		build_logger.error(process.stderr.decode('utf-8'))
 
-def create_logger (args):
-	log_format = "%(asctime)s - %(levelname)s: %(message)s"
-	date_format = '%d-%m-%Y %H:%M:%S'
-	if args.debug:
-		logging.basicConfig(level=logging.DEBUG, format=log_format, datefmt=date_format)
-	elif args.verbose:
-		logging.basicConfig(level=logging.INFO, format=log_format, datefmt=date_format)
-	else:
-		logging.basicConfig(level=logging.WARNING, format=log_format, datefmt=date_format)
+def install_python_dependencies():
+	pip_call = ['sudo', 'pip3', 'install', 'pymysql']
+	try:
+		subprocess.run(pip_call, check=True)
+	except subprocess.CalledProcessError as err:
+		build_logger.error('Could not install python dependencies.')
+		print_process_error(err)
+		exit(1)
 
-def setup_packages (args):
+def setup_packages ():
 	if not _is_architecture_supported():
-		logging.error('Your system system architecture is not supported.')
+		build_logger.error('Your system system architecture is not supported.')
 		exit (1)
 	else:
 		my_architecture = _get_system_architecture()
 		list_of_packages_needed = _build_list_of_packages(my_architecture)
-		_install_packages(args, list_of_packages_needed)
+		_install_packages(list_of_packages_needed)
 
-def make_c_files (args):
-	logging.info('Making C files...')
+def make_c_files ():
+	build_logger.info('Making C files...')
 	if args.clean:
 		_call_std_subprocess(['make', '--quiet', 'clean'])
 	elif not args.clean:
 		_call_std_subprocess(['make', '--quiet'])
 	else:
-		logging.info('C files built up!')
+		build_logger.info('C files built up!')
 
 def setup_env ():
 	# Check if variable is not already set
@@ -141,58 +150,69 @@ def setup_env ():
 	local_env = 'export PYTHONPATH=' + dirname
 	bashrc_env = 'echo \"export PYTHONPATH=' + dirname + '\" >> ~/.bashrc'
 	try:
-		subprocess.run(local_env, shell=True)
+		subprocess.run(local_env, shell=True, check=True)
 	except subprocess.CalledProcessError as err:
-		logging.error('Could not setup environment.')
+		build_logger.error('Could not setup environment.')
 		print_process_error(err)
 		exit(1)
 	try:
-		subprocess.run(bashrc_env, shell=True)
+		subprocess.run(bashrc_env, shell=True, check=True)
 	except subprocess.CalledProcessError as err:
-		logging.error('Could not setup environment.')
+		build_logger.error('Could not setup environment.')
 		print_process_error(err)
 		exit(1)
 
 
 def install_data_base ():
-	logging.info('Setting up database as root, please enter your new password.\n')
-
+	print('Setting up database for root.\n')
 	try:
-		password = getpass.getpass()
+		password = getpass.getpass('New password: ')
 	except KeyboardInterrupt:
+		build_logger.warning('Exiting script')
 		interrupt_handler(sys.exc_info())
 
-	start_mysql = ['sudo', '/etc/init.d/mysql', 'start']
 	try:
+		start_mysql = ['sudo', '/etc/init.d/mysql', 'start']
 		subprocess.run(start_mysql)
 	except subprocess.CalledProcessError as err:
-		logging.error('Could not start MYSQL.')
+		build_logger.error('Could not start MYSQL.')
 		print_process_error(err)
 		exit(1)
 
-	SQL_querry = 'DROP USER \'root\'@\'localhost\'; CREATE USER \'root\'@\'localhost\' IDENTIFIED BY \'' + password + '\'; ' + 'GRANT ALL PRIVILEGES ON *.* TO \'root\'@\'localhost\' WITH GRANT OPTION; ' + 'FLUSH PRIVILEGES;'
-	root_cmd = 'sudo su -c \"/bin/sh\"'
-	mysql_cmd = 'mysql -u root -e \"' + SQL_querry + '\"'
-	with subprocess.Popen (root_cmd, stdin=subprocess.PIPE, stdout=subprocess.PIPE, shell=True) as root:
-		root.stdin.write(mysql_cmd.encode())
-		root.stdin.close()
-		root.wait()
+	try:
+		check_user_config = ['mysql', '-N', '-B', '-u', 'root', '-p' + password, '-e', '\"SELECT EXISTS(SELECT * FROM mysql.user WHERE user = \'root\');\"']
+		subprocess.run(' '.join(check_user_config), stderr=subprocess.PIPE, stdout=subprocess.PIPE, check=True, shell=True)
+	except subprocess.CalledProcessError:
+		user_is_already_configured = False
+	else:
+		user_is_already_configured = True
+
+	if not user_is_already_configured:
+		SQL_querry = 'DROP USER \'root\'@\'localhost\'; CREATE USER \'root\'@\'localhost\' IDENTIFIED BY \'' + password + '\'; ' + 'GRANT ALL PRIVILEGES ON *.* TO \'root\'@\'localhost\' WITH GRANT OPTION; ' + 'FLUSH PRIVILEGES;'
+		root_cmd = 'sudo su -c \"/bin/sh\"'
+		mysql_cmd = 'mysql -u root -e \"' + SQL_querry + '\"'
+		with subprocess.Popen (root_cmd, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True) as root:
+			root.stdin.write(mysql_cmd.encode())
+			root.stdin.close()
+			root.wait()
 
 	SQL_querry = '../database_setup/setup.sql'
 	mysql_cmd = 'mysql -u root -p' + password + ' < ' + SQL_querry
 	try:
-		setup = subprocess.run(mysql_cmd, stdin=subprocess.PIPE, stdout=subprocess.PIPE, shell=True)
+		setup = subprocess.run(mysql_cmd, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
 	except subprocess.CalledProcessError as err:
 		print_process_error(err)
-	logging.info(setup.stdout.decode('utf-8'))
+	build_logger.info(setup.stdout.decode('utf-8'))
+
+args = arg_parser()
+build_logger = create_logger()
 
 def main ():
-	args = arg_parser()
-	create_logger(args)
-	setup_packages(args)
-	make_c_files(args)
+	setup_packages()
+	make_c_files()
 	install_data_base()
 	setup_env()
+	install_python_dependencies()
 	
 if __name__ == '__main__':
 	main()
